+++ source/client/ftlib.c
@@ -35,7 +35,7 @@ static void CL_FTLibModule_Print( const char *msg )
 	Com_Printf( "%s", msg );
 }
 
-static void *CL_FTLibModule_MemAlloc( mempool_t *pool, int size, const char *filename, int fileline )
+static void *CL_FTLibModule_MemAlloc( mempool_t *pool, size_t size, const char *filename, int fileline )
 {
 	return _Mem_Alloc( pool, size, MEMPOOL_CINMODULE, 0, filename, fileline );
 }
