+++ source/qcommon/ascript.c
@@ -34,7 +34,7 @@ static void Com_ScriptModule_Print( const char *msg )
 	Com_Printf( "%s", msg );
 }
 
-static void *Com_ScriptModule_MemAlloc( mempool_t *pool, int size, const char *filename, int fileline )
+static void *Com_ScriptModule_MemAlloc( mempool_t *pool, size_t size, const char *filename, int fileline )
 {
 	return _Mem_Alloc( pool, size, MEMPOOL_ANGELSCRIPT, 0, filename, fileline );
 }
